[Add] App clip setting and app logos
