+++ RoomEscape/GBZGClip/ContentView.swift
@@ -1,0 +1,21 @@
+//
+//  ContentView.swift
+//  GBZGClip
+//
+//  Created by Hyeon-sang Lee on 2022/07/27.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        Text("Hello, world!")
+            .padding()
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
