+++ RoomEscape/GBZGClip/GBZGClip.entitlements
@@ -4,7 +4,7 @@
 <dict>
 	<key>com.apple.developer.associated-domains</key>
 	<array>
-		<string>appclips:developeracademy-postech.github.io/MC3_TEAM1_LATER/</string>
+		<string>appclips:mc3-later-gbzg.web.app</string>
 	</array>
 	<key>com.apple.developer.parent-application-identifiers</key>
 	<array>
